+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 * Pagination for expected and received pieces. Refs UIREC-213.
 * refactor psets away from backend '.all' permissions. Refs UIREC-177.
 * Filters for expected and received pieces. Refs UIREC-191.
+* Item status changes unclear from receiving screen. Refs UIREC-218.
+* Item status unclear in receiving screen. Refs UIREC-214.
 
 ## [2.0.3](https://github.com/folio-org/ui-receiving/tree/v2.0.3) (2021-12-08)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v2.0.2...v2.0.3)